+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - implement ExtendedColour trait for Brown and add tests for name_colour function(pr [#65])
 - implement ExtendedColour trait for Cyan(pr [#66])
 - implement ExtendedColour trait and add comprehensive tests for Green(pr [#67])
+- implement ExtendedColour trait and add comprehensive tests for Red(pr [#68])
 
 ### Changed
 
@@ -161,6 +162,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#65]: https://github.com/jerus-org/named-colour/pull/65
 [#66]: https://github.com/jerus-org/named-colour/pull/66
 [#67]: https://github.com/jerus-org/named-colour/pull/67
+[#68]: https://github.com/jerus-org/named-colour/pull/68
 [Unreleased]: https://github.com/jerus-org/named-colour/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/jerus-org/named-colour/compare/v0.1.6...v0.2.0
 [0.1.6]: https://github.com/jerus-org/named-colour/compare/v0.1.5...v0.1.6